Reworked color an dicon caching

+++ Demo-Swift/slds-sample-app/library/model/IconObject.swift
@@ -19,7 +19,7 @@ enum IconObjectType : String {
 struct IconObject {
     var type : IconObjectType
     var index: NSInteger
-    var size : Int16 = 0
+    var size : CGFloat = 0.0
     
     //––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
     
@@ -53,22 +53,22 @@ struct IconObject {
         switch self.type {
         case .action :
             if let value = SLDSActionIconType.init(rawValue: self.index) {
-                return UIImage.sldsActionIcon(value, withSize: self.size)
+                return UIImage.sldsActionIcon(value, withSize: Int16(self.size))
             }
             
         case .custom :
             if let value = SLDSCustomIconType.init(rawValue: self.index) {
-                return UIImage.sldsCustomIcon(value, withSize: self.size)
+                return UIImage.sldsCustomIcon(value, withSize: Int16(self.size))
             }
             
         case .standard :
             if let value = SLDSStandIconType.init(rawValue: index) {
-                return UIImage.sldsStandardIcon(value, withSize: self.size)
+                return UIImage.sldsStandardIcon(value, withSize: Int16(self.size))
             }
             
         case .utility :
             if let value = SLDSUtilityIconType.init(rawValue: self.index) {
-                return UIImage.sldsUtilityIcon(value, withSize: self.size)
+                return UIImage.sldsUtilityIcon(value, withSize: Int16(self.size))
             }
         }
